[Refactor] nlohmann-jsonのヘッダを移動
他のサードパーティ製ライブラリと同様になるよう、nlohmann-jsonのヘッダを
設置するディレクトリを src/external-lib/include から
src/external-lib/include/nlohmann へ移動する。
