+++ src/external-lib/include-json.h
@@ -6,7 +6,7 @@
 #pragma warning(push, 0)
 #endif
 
-#include <json.hpp>
+#include <nlohmann/json.hpp>
 
 #if defined(_MSC_VER)
 #pragma warning(pop)
